command_hub/requirements.txt
Flask==3.0.2
azure-iot-hub==2.6.1



